# swagger/aos3.yaml
openapi: 3.0.0
info: 
  title: Vigilion API
  description: API per la web app Vigilion
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Localhost server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - id
        - firstname
        - lastname
        - email
        - password
        - district
      properties:
        id:
          type: string
          description: UUID dell'utente
        firstname:
          type: string
          description: Nome dell'utente
        lastname:
          type: string
          description: Cognome dell'utente
        email:
          type: string
          description: Email dell'utente
        password:
          type: string
          description: Password dell'utente
        district:
          type: string
          description: Circoscrizione di residenza dell'utente
        notifications:
          type: array
          description: Array contenente le circoscrizioni da cui l'utente riceverà notifiche
        warned:
          type: boolean
          description: Valore booleano che indica se l'utente è stato ammonito per l'utilizzo della piattaforma
        blocked:
          type: boolean
          description: Valore booleano che indica se l'utente è stato bandito dalla creazione di segnalazioni
        admin:
          type: boolean
          description: Valore booleano che indica se l'utente è amministratore
        feedbacks:
          type: array
          description: Lista dei feedback inviati dall'utente
          items:
            type: object
            properties:
              text:
                type: string
                description: Testo del feedback
              date:
                type: string
                description: Data di invio del feedback (formato yyyy-mm-dd)
    Report:
      type: object
      required:
        - user
        - typology
        - notes
        - location
        - district
      properties:
        id:
          type: integer
          description: ID autogenerato incrementale della segnalazione
        user:
          type: string
          description: UUID dell'utente che ha creato la segnalazione
        typology:
          type: string
          enum: [Furto, Aggressione, Molestia, Soggetto armato, Soggetto alterato, Altro]
          description: Tipologia della segnalazione
        notes:
          type: string
          description: Descrizione della segnalazione
        location:
          type: string
          description: Posizione della segnalazione (latitudine/longitudine)
        upvote:
          type: integer
          description: Numero di upvote ricevuti
        downvote:
          type: integer
          description: Numero di downvote ricevuti
        createdtime:
          type: string
          description: Data e ora di creazione della segnalazione (formato yyyy-mm-dd HH:ii)
...
  responses:
    Created:
      description: Risorsa creata
    NotFound:
      description: Risorsa non trovata
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Risorsa non trovata
    BadRequest:
      description: Richiesta non valida
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Parametri non validi
    Unauthorized:
      description: L'utente non è autorizzato a compiere questa azione
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Utente non autorizzato
    ServerError:
      description: Errore del server
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Tipologia di errore
              details:
                type: string
                description: Dettagli dell'errore
paths:
  /authentication:
    post:
      summary: Autenticazione dell'utente
      description: Autentica l'utente e ritorna il token JWT in un cookie apposito
      requestBody:
        description: Credenziali per l'autenticazione (email e password)
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  description: Email dell'utente
                password:
                  type: string
                  description: Password dell'utente
      responses:
        "200":
          description: Token creato e salvato nei cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: True se l'autenticazione è riuscita, false altrimenti
                  token:
                    type: string
                    description: Token JWT per l'autenticazione
        "404":
          description: Utente non trovato con le credenziali fornite
          $ref: '#/components/responses/NotFound'
...
  /authentication/check:
    get:
      summary: Verifica autenticazione
      description: Verifica se l'utente è autenticato controllando il cookie di autenticazione
      responses:
        "200":
          description: Stato dell'autenticazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                    description: True se l'utente è autenticato
                  user:
                    type: object
                    properties:
                      email:
                        type: string
                        description: Email dell'utente autenticato
                      id:
                        type: string
                        description: ID dell'utente autenticato
        "400":
          description: Token non trovato nei cookie
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Errore nella verifica del token
          $ref: "#/components/responses/Unauthorized"
  /districts:
    get:
      summary: Espone il file circoscrizioni.geojson alle pagine frontend
      responses:
        "200":
          description: File contenente i poligoni che rappresentano le circoscrizioni del comune di Trento
          content:
            application/geo+json:
              schema:
                type: array
        "404":
          description: File non trovato
          $ref: "#/components/responses/NotFound"
...
  /users:
    summary: Rappresentazione della lista di utenti
    description: In chiamata GET ritorna la lista di tutti gli utenti registrati, in POST crea un nuovo utente
    get:
      summary: Lista degli utenti
      responses:
        "200":
          description: Lista degli utenti
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "404":
          description: Nessun utente trovato
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      summary: Crea un nuovo utente
      requestBody:
        description: Utente da creare
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Utente creato
          $ref: "#/components/responses/Created"
        "400":
          description: Email o password non valide, oppure email già utilizzata
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /reports:
    summary: Gestione delle segnalazioni
    description: Creazione in POST di una nuova segnalazione e restituzione in GET di tutte le segnalazioni
    post:
      summary: Creazione nuova segnalazione
      security:
        - bearerAuth: []
      requestBody:
        description: Segnalazione da creare
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Report"
      responses:
        "201":
          $ref: "#/components/responses/Created"
          description: Segnalazione creata
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: La segnalazione non appartiene ad alcuna circoscrizione
    get:
      summary: Lista delle segnalazioni
      responses:
        "200":
          description: Lista delle segnalazioni
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "404":
          description: Nessuna segnalazione trovata
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...